refactor: extract chunked download into Client
- Move chunked download logic from OtaClient to Client::download
- Add ChunkedDownloadError and 5 chunking constants to http module
- OtaClient now delegates to GithubClient::download with auth closures
- Enables reuse of chunked download for non-OTA purposes
